Release v1.1.0.4.
This will be the latest version until three or four years later because the developers are about to take the high school entrance examination. They will also experience a busy high school life. So, it's difficult for them to find time to continue updating this project during this period. However, we promise that the Issues raised during this period will never be ignored!